+++ scvi/benchmark.py
@@ -4,9 +4,8 @@ import torch
 from torch.utils.data import DataLoader
 from torch.utils.data.sampler import SubsetRandomSampler
 
-from scvi.dataset import CortexDataset
+from scvi.dataset import CortexDataset, BrainLargeDataset
 from scvi.dataset.utils import get_data_loaders, get_raw_data
-from scvi.dataset.benchmark_hyperparameters import benchmark_hyperparameters
 from scvi.metrics.adapt_encoder import adapt_encoder
 from scvi.metrics.classification import compute_accuracy_rf, compute_accuracy_svc
 from scvi.metrics.clustering import entropy_batch_mixing, get_latent
@@ -16,6 +15,24 @@ from scvi.metrics.visualization import show_t_sne
 from scvi.models import VAE, SVAEC
 from scvi.models.classifier import Classifier
 from scvi.train import train, train_classifier, train_semi_supervised_jointly, train_semi_supervised_alternately
+
+
+def benchmark_hyperparameters(gene_dataset):
+    """
+    Possibly specify hyper parameters according to Table2 in "Bayesian Inference for a Generative Model of
+    transcriptome Profiles from Single-cell RNA Sequencing"
+    :param gene_dataset:
+    :return:
+    """
+    hyperparameters = dict()
+    if isinstance(gene_dataset, BrainLargeDataset):
+        if gene_dataset.nb_genes > 10000 and gene_dataset.nb_genes < 15000:
+            hyperparameters["n_layers"] = 2
+        elif gene_dataset.nb_genes > 15000:
+            hyperparameters["n_layers"] = 3
+            hyperparameters["n_hidden"] = 256
+
+    return hyperparameters
 
 
 def run_benchmarks(gene_dataset, model=VAE, n_epochs=1000, lr=1e-3, use_batches=False, use_cuda=True,
